Attach sidebar to gui.left
gui.left is a vertical flow, mod_gui is horizontal. By attaching to
gui.left, the sidebar windows are guaranteed to stick to the left side
of the screen, even if other mods are present.
Additionally, on_configuration change had to be modified to remove old
gui elements attached to mod_gui.